Add PluginContext.customDragAndDropHandlers
